Refactor detail timeline library
