+++ apps/web/src/components/detail/lib/timeline/labels.ts
@@ -1,0 +1,108 @@
+/**
+ * Maps event-kind strings (`agent.run-started`, `state.transitioned`, ...) to
+ * human-readable labels. This is only a label registry; timeline subscription
+ * and visibility are governed by @goose-hub/core/event-stream/issue-timeline.
+ *
+ * NOTE: this is event-kind labelling. `STATE_LABEL` in `@/lib/constants`
+ * separately maps `factory:*` work-item state names; keep them apart.
+ */
+export const EVENT_KIND_LABEL: Record<string, string> = {
+  'state.transitioned': 'State transitioned',
+  'milestone.activated': 'Milestone activated',
+  'agent.spawned': 'Agent spawned',
+  'agent.decision-summary': 'Decision summary',
+  'agent.decision-summary-live': 'Decision (live)',
+  'agent.disclosure': 'Input summarized',
+  'agent.terminated': 'Agent terminated',
+  'agent.log': 'Agent log',
+  'gate.awaiting-human': 'Gate — awaiting human',
+  'system.note': 'Note',
+  'manual.action': 'Manual action',
+  'agent.run-started': 'Agent run started',
+  'agent.run-completed': 'Agent run completed',
+  'agent.run-failed': 'Agent run failed',
+  'agent.budget-exceeded': 'Agent budget exceeded',
+  'agent.tool-call': 'Tool call',
+  'agent.tool-result': 'Tool result (error)',
+  'tool.stdout-truncated': 'Stdout truncated',
+  'tool.timeout': 'Timeout',
+  'agent.fallback-triggered': 'Fallback triggered',
+  'agent.triage-complete': 'Triage complete',
+  'agent.investigation-complete': 'Investigation complete',
+  'agent.investigation-context-injected': 'Investigation context injected',
+  'agent.related-surface-manifest-created': 'Related surface manifest',
+  'agent.discovery-budget-exceeded': 'Discovery budget exceeded',
+  'agent.wrong-surface-guard': 'Wrong surface guard',
+  'agent.path-normalized': 'Path normalized',
+  'agent.output-repaired': 'Output repaired',
+  'agent.output-repair-failed': 'Output repair failed',
+  'agent.output-fact-mismatch': 'Output fact mismatch',
+  'agent.contract-gate-blocked': 'Contract gate blocked',
+  'acceptance.contract-authored': 'Acceptance contract authored',
+  'symbol-index.lookup': 'Symbol index lookup',
+  'symbol-index.hints-used': 'Symbol hints used',
+  'agent.implement-complete': 'Implement complete',
+  'agent.fix-feedback-complete': 'Fix feedback complete',
+  'agent.retry-escalated': 'Retry escalated',
+  'pr.opened': 'PR opened',
+  'pr.merged': 'PR merged',
+  'gate.approved': 'Gate approved',
+  'gate.rejected': 'Gate rejected',
+  'review.completed': 'Review completed',
+  'review.wave-completed': 'Review wave completed',
+  'review.wave-failed': 'Review wave failed',
+  'review.converged': 'Review converged',
+  'review.escalated': 'Review escalated',
+  'review.slot-completed': 'Review slot completed',
+  'evidence.no-spec-declared': 'Evidence — no spec declared',
+  'evidence.playwright-repro-skipped': 'Playwright repro skipped',
+  'evidence.playwright-ran': 'Playwright evidence ran',
+  'evidence.posted': 'Evidence posted',
+  'evidence.post-skipped': 'Evidence post skipped',
+  'evidence.post-failed': 'Evidence post failed',
+  'agent.verify-command': 'Verify command',
+  'qa.completed': 'QA completed',
+  'qa.structural-passed': 'QA structural passed',
+  'qa.structural-failed': 'QA structural failed',
+  'qa.functional-passed': 'QA functional passed',
+  'qa.functional-failed': 'QA functional failed',
+  'qa.regression-passed': 'QA regression passed',
+  'qa.regression-failed': 'QA regression failed',
+  'qa.verification-summary-built': 'QA verification summary built',
+  'qa.verification-blocked': 'QA verification blocked',
+  'retrospective.completed': 'Retrospective completed',
+  'grill.question-posted': 'Grill question posted',
+  'grill.completed': 'Grill completed',
+  'grill.decision-crystallized': 'Grill decision crystallized',
+  'prd.drafted': 'PRD drafted',
+  'prd.advisor-skipped': 'Advisor skipped',
+  'prd.approved': 'PRD approved',
+  'prd.rejected': 'PRD rejected',
+  'prd.revised': 'PRD revision requested',
+  'prd.declined': 'Feature declined',
+  'decompose.completed': 'Decompose complete',
+  'spec.completed': 'Spec authored',
+  'agent.model-selected': 'Model selected',
+  'swarm.heartbeat': 'Swarm heartbeat',
+  'swarm.scout-completed': 'Scout completed',
+  'swarm.scout-failed': 'Scout failed',
+  'swarm.scout-timeout': 'Scout timeout',
+  'swarm.wave-completed': 'Wave completed',
+  'swarm.wave-halted': 'Wave halted',
+  'swarm.wave-incomplete': 'Wave incomplete',
+  'parallel-implement.iteration-started': 'Parallel iteration started',
+  'parallel-implement.wp-started': 'Work package started',
+  'parallel-implement.wp-committed': 'Work package committed',
+  'parallel-implement.wp-failed': 'Work package failed',
+  'parallel-implement.wp-timeout': 'Work package timed out',
+  'parallel-implement.wp-commit-failed': 'Work package commit failed',
+  'parallel-implement.exhausted': 'Parallel implement exhausted',
+  'dev-review.started': 'Dev review started',
+  'dev-review.completed': 'Dev review completed',
+  'dev-review.failed': 'Dev review failed',
+  'dev-review.error': 'Dev review error',
+  'dev-review.budget-skipped': 'Dev review budget skipped',
+  'dev-review.response-started': 'Dev review response started',
+  'dev-review.response-completed': 'Dev review response completed',
+  'dev-review.response-failed': 'Dev review response failed',
+};
